refactor: move brokerage logic
